chore(experiments): moving data collection component to the legacy folder.
